JSON configuration, cosmetic changes in console

+++ src/main/java/org/rathercruel/bot/commands/JoinToCreate.java
@@ -18,42 +18,36 @@ import java.util.List;
  */
 public class JoinToCreate extends ListenerAdapter {
 
-    private int vcCreatedInt = 0;
-
     @Override
     public void onGuildVoiceUpdate(GuildVoiceUpdateEvent event) {
-        AudioChannel audioChannel = event.getChannelJoined();
-        Guild guild = event.getGuild();
-        Member member = event.getMember();
+        if (BotConfiguration.joinToCreateBool) {
+            AudioChannel audioChannel = event.getChannelJoined();
+            Guild guild = event.getGuild();
+            Member member = event.getMember();
 
-        if (audioChannel != null) {
-//            vcCreatedInt++;
-            long channelID = audioChannel.getIdLong();
-            List<Long> allow = new ArrayList<>();
-            allow.add(Permission.ALL_CHANNEL_PERMISSIONS);
-            long test = Permission.ALL_CHANNEL_PERMISSIONS;
-            long deny = 0;
+            if (audioChannel != null) {
+                long channelID = audioChannel.getIdLong();
+                List<Long> allow = new ArrayList<>();
+                allow.add(Permission.ALL_CHANNEL_PERMISSIONS);
+                long test = Permission.ALL_CHANNEL_PERMISSIONS;
+                long deny = 0;
 
-            if (channelID == BotConfiguration.joinToCreateChannelID) {
-                audioChannel.createCopy().setName("[\uD83E\uDD1D] " + member.getEffectiveName() + "'s Voice").queue((newChannel) -> {
-                    guild.moveVoiceMember(member, (AudioChannel) newChannel).queue();
-                    ((AudioChannel) newChannel).getManager().putMemberPermissionOverride(member.getIdLong(), test, deny).queue();
-                });
-            }
-        } else {
-//            vcCreatedInt--;
-
-            // REMOVE EMPTY VC
-            Category category = event.getGuild().getCategoryById(BotConfiguration.joinToCreateCategoryID);
-            List<VoiceChannel> vcList = category.getVoiceChannels();
-            for (int i = 1; i < vcList.size(); i++) {
-                if (vcList.get(i).getMembers().size() == 0) {
-                    long tempID = vcList.get(i).getIdLong();
-                    guild.getVoiceChannelById(tempID).delete().queue();
+                if (channelID == BotConfiguration.joinToCreateChannelID) {
+                    audioChannel.createCopy().setName("[\uD83E\uDD1D] " + member.getEffectiveName() + "'s Voice").queue((newChannel) -> {
+                        guild.moveVoiceMember(member, (AudioChannel) newChannel).queue();
+                        ((AudioChannel) newChannel).getManager().putMemberPermissionOverride(member.getIdLong(), test, deny).queue();
+                    });
+                }
+            } else {
+                Category category = event.getGuild().getCategoryById(BotConfiguration.joinToCreateCategoryID);
+                List<VoiceChannel> vcList = category.getVoiceChannels();
+                for (int i = 1; i < vcList.size(); i++) {
+                    if (vcList.get(i).getMembers().size() == 0) {
+                        long tempID = vcList.get(i).getIdLong();
+                        guild.getVoiceChannelById(tempID).delete().queue();
+                    }
                 }
             }
         }
-//        guild.getVoiceChannelById(BotConfiguration.guildVoiceChannelCount)
-//                .getManager().setName("\uD83C\uDF3F VC created: " + vcCreatedInt).queue();
     }
 }
